robot-register-proxy/requirements.txt
paho-mqtt==2.0.0
PyJWT==2.8.0
PyYAML==6.0.1
